requirements.txt
certifi==2021.5.30
chardet==4.0.0
et-xmlfile==1.1.0
idna==2.10
openpyxl==3.0.7
pkg-resources==0.0.0
python-dotenv==0.18.0
requests==2.25.1
urllib3==1.26.6
